pyproject.toml
[tool.poetry]
name = "pylibretro"
version = "0.1.0"
description = "WIP Python library that runs Libretro cores"
authors = ["James Ravindran"]
license = "AGPL-3.0-or-later"
readme = "README.md"
keywords = ["libretro", "emulator", "retroarch"]
repository = "https://github.com/jamesravi/pylibretro"
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Operating System :: POSIX :: Linux",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.2.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
toml = "^0.10.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
